+++ src/modules/osubot/middleware.js
@@ -1,18 +1,20 @@
-import yaml from 'js-yaml'
+import { safeLoad } from 'js-yaml'
 import { readFileSync } from 'fs'
 import chalk from 'chalk'
+import { mod } from '../../core/log'
+import analyzer from '../../core/analyzer'
 const log = {
     private: {},
     group: {},
 }
-const { repeater: config } = yaml.safeLoad(readFileSync('config.yml')).osubot
+const { repeater: config } = safeLoad(readFileSync('config.yml')).osubot
 /**
  * This middleware counts the repeated times of a message
  * if it reaches 3, the bot repeats it
  * @param {Message} msg 
  */
 export default function repeater(msg) {
-    if (/!|！/.test(msg.param.message) === false && 
+    if (/^[!！]/.test(msg.param.message) === false && 
         msg.type === 'group') {
         if ('notAllowed' in config && config.notAllowed.includes(msg.target)) return
         if ('allowed' in config && !config.allowed.includes(msg.target)) return
@@ -22,8 +24,9 @@ export default function repeater(msg) {
         if (log[msg.type][msg.target].count === config.times) {
             const timeout = Math.round(Math.random() * 200000),
                   repeatTarget = log[msg.type][msg.target]
-            console.log(`${chalk.blue('[MOD]')} ${chalk.gray(new Date().toString())} by osubot middleware: attempting to repeat \`${msg.param.message}' in ${msg.type === 'group' ? `${msg.type} ${msg.target}` : chalk.yellow(`${msg.type} ${msg.target}`)} in ${Math.round(timeout / 1000)} secs`)
+            mod('osubot middleware', `attempting to repeat \`${msg.param.message}' in ${msg.type === 'group' ? `${msg.type} ${msg.target}` : chalk.yellow(`${msg.type} ${msg.target}`)} in ${Math.round(timeout / 1000)} secs`)
             setTimeout(() => { msg.send(repeatTarget.message) }, timeout)
+            analyzer(msg, 'middleware', 'osubotRepeat')
             delete log[msg.type][msg.target]
         } 
     }